Rename files called uts46.rs to deprecated.rs
