+++ idna/tests/tests.rs
@@ -1,7 +1,7 @@
 use tester as test;
 
 mod punycode;
-mod uts46;
+mod deprecated;
 mod uts46bis;
 
 fn main() {
@@ -20,7 +20,7 @@ fn main() {
             })
         };
         punycode::collect_tests(&mut add_test);
-        uts46::collect_tests(&mut add_test);
+        deprecated::collect_tests(&mut add_test);
         uts46bis::collect_tests(&mut add_test);
     }
     test::test_main(&std::env::args().collect::<Vec<_>>(), tests, None)
